requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
asyncpg==0.27.0
fastapi==0.89.1
ormar==0.12.1
psycopg2-binary==2.9.6
SQLAlchemy~=1.4.41
pydantic~=1.10.4
databases~=0.6.2
redis~=4.6.0
python-dotenv~=1.0.0
alembic~=1.11.1